Change the config file format and download path.
The download path format now allows adding a type, which is used to determine
which kind of API will be used, instead of using the name of the section. This
allows people to specify multiple instances of Gitlab to be backed up together.

Also, updated README with instructions to use `python setup.py install` instead
of using `pip install` to download only the required package.